requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.1.0
django-filter>=23.5
django-extensions>=3.2.3
cryptography>=41.0.0 